requirements.txt
Flask
pymongo
python-dotenv
flask-cors
requests
google-cloud-speech
google-auth
google-generativeai